+++ src/main/java/org/jabref/gui/shared/PullChangesFromSharedAction.java
@@ -3,6 +3,7 @@ package org.jabref.gui.shared;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionHelper;
 import org.jabref.gui.actions.SimpleCommand;
+import org.jabref.logic.shared.DatabaseLocation;
 import org.jabref.logic.shared.DatabaseSynchronizer;
 
 public class PullChangesFromSharedAction extends SimpleCommand {
@@ -16,7 +17,9 @@ public class PullChangesFromSharedAction extends SimpleCommand {
     }
 
     public void execute() {
-        stateManager.getActiveDatabase().ifPresent(databaseContext -> {
+        stateManager.getActiveDatabase()
+                    .filter(databaseContext->databaseContext.getLocation() == DatabaseLocation.SHARED)
+                    .ifPresent(databaseContext -> {
             DatabaseSynchronizer dbmsSynchronizer = databaseContext.getDBMSSynchronizer();
             dbmsSynchronizer.pullChanges();
         });